[Refactor] #21 - 네트워크 에러 중첩 구조로 변경
